refactor(reader): extract ContinuousWindowController from ContinuousReaderView (#390)
Move the sliding-window state machine — topIndex, webViews, measuredHeights,
pendingInitialScroll, landing-hold, reapply-after-fallback, window shifts,
renderer recovery, the recycled-WebView pool, and the append/prepend/remove
chapter operations — into a new ContinuousWindowController. The controller
talks to NestedScrollView-flavour scroll primitives via a ContinuousScrollPort
interface, so the algorithm is decoupled from the concrete View.

ContinuousReaderView retains the NestedScrollView-specific pieces that must
live on a real View: touch/fling arbitration, the requestChildFocus / scrollBy
overrides (selection-jump fix), the fling cap, and computeScroll (delegates to
controller.tickLandingHold each frame). It also owns the selection-active
counter used by touch overrides.

Public API is unchanged — annotationsAvailable, readaloudAvailable, and
onPlayFromHereSelection are now delegated properties; initialize / navigateTo /
updatePreferences / applyAnnotationHighlights / applySearchHighlights /
highlightInChapter / clearHighlightInChapter / scrollByPage all delegate to
the controller.

ContinuousReaderView.kt shrinks from 1196 to ~230 lines.

Follow-ups (per issue #390's 3-4 PR plan): abstract ChapterWebView construction
behind a factory so openWindowAt / pendingInitialScroll timing bugs become
JVM-testable in isolation.

Refs #390
